+++ CHANGELOG.md
@@ -51,3 +51,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove `setDefaultOptions()`, in favor of simply returning object on
   `beforeSend()`.
+