+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed
+
+- `localBaseURL` option removed
+
+### Fixed
+
+- Fixed typo in README on array/object as value for key.
+- Fixed typo in docs where it said to be using `beforeSend()` return instead of
+`dynamicOptions()`.
+
 ## [2.0.0] - 2022-01-16
 
 ### Added