+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.1] - 2022-01-17
 
 ### Removed
 
@@ -85,3 +85,6 @@ options as object parameters.
 - Remove `setDefaultOptions()`, in favor of simply returning object on
   `beforeSend()`.
   
+
+[Unreleased]: https://github.com/D-Marc1/far-fetch/compare/v2.0.1...HEAD
+[2.0.1]: https://github.com/D-Marc1/far-fetch/compare/v2.0.0...v2.0.1